feat(novelEditorAi): 实现章节内容生成和管理功能
- 新增章节内容生成接口和相关模型
- 实现章节内容生成服务，包括构建章节上下文、生成提示词和调用AI模型
- 添加流式生成和保存章节内容的功能
- 优化日志配置，使用 novelEditorAi.mapper 包名

+++ novel-main/src/main/java/com/soukon/novelEditorAi/controller/CharacterController.java
@@ -25,7 +25,7 @@ public class CharacterController {
     }
     
     @GetMapping("/project/{projectId}")
-    public Result<List<Character>> listByProjectId(@PathVariable("projectId") Integer projectId) {
+    public Result<List<Character>> listByProjectId(@PathVariable("projectId") Long projectId) {
         LambdaQueryWrapper<Character> queryWrapper = new LambdaQueryWrapper<>();
         queryWrapper.eq(Character::getProjectId, projectId);
         queryWrapper.orderByAsc(Character::getName);
